requirements.txt
requests==2.9.0
Pillow==3.3.0
tweepy==3.5.0